+++ src/c/regmaps/C0microSDPlus/top_regs.c
@@ -7,26 +7,26 @@
 
 
 uint32_t
-getTopFlashOtpEntry(uint32_t offset)
+getTopFlashOtpEntry(size_t offset)
 {
 	return *(volatile uint32_t *)(0x8000000 + offset);
 }
 
 void
-setTopFlashOtpEntry(uint32_t offset, uint32_t *  value)
+setTopFlashOtpEntry(size_t offset, uint32_t *  value)
 {
 	*(volatile uint32_t *)(0x8000000 + offset) = *value;
 }
 
 
 uint32_t
-getTopLramEntry(uint32_t offset)
+getTopLramEntry(size_t offset)
 {
 	return *(volatile uint32_t *)(0x8100000 + offset);
 }
 
 void
-setTopLramEntry(uint32_t offset, uint32_t *  value)
+setTopLramEntry(size_t offset, uint32_t *  value)
 {
 	*(volatile uint32_t *)(0x8100000 + offset) = *value;
 }
@@ -34,13 +34,13 @@ setTopLramEntry(uint32_t offset, uint32_t *  value)
 
 
 uint32_t
-getTopIoBuffEntry(uint32_t offset)
+getTopIoBuffEntry(size_t offset)
 {
 	return *(volatile uint32_t *)(0x8300000 + offset);
 }
 
 void
-setTopIoBuffEntry(uint32_t offset, uint32_t *  value)
+setTopIoBuffEntry(size_t offset, uint32_t *  value)
 {
 	*(volatile uint32_t *)(0x8300000 + offset) = *value;
 }
